+++ mayan/apps/document_indexing/apps.py
@@ -34,8 +34,9 @@ from .links import (
 from .licenses import *  # NOQA
 from .permissions import (
     permission_document_indexing_create, permission_document_indexing_delete,
-    permission_document_indexing_edit, permission_document_indexing_rebuild,
-    permission_document_indexing_view
+    permission_document_indexing_edit,
+    permission_document_indexing_instance_view,
+    permission_document_indexing_rebuild, permission_document_indexing_view
 )
 from .queues import *  # NOQA
 from .widgets import get_instance_link, index_instance_item_link, node_level
@@ -73,6 +74,7 @@ class DocumentIndexingApp(MayanAppConfig):
                 permission_document_indexing_create,
                 permission_document_indexing_delete,
                 permission_document_indexing_edit,
+                permission_document_indexing_instance_view,
                 permission_document_indexing_rebuild,
                 permission_document_indexing_view,
             )